requirements.txt
flask==3.0.2
flask-cors==4.0.0
bcrypt==4.1.2
python-dotenv==1.0.1
pyjwt==2.8.0
boto3==1.34.91
